+++ common/src/main/java/name/dashkal/minecraft/hexresearch/network/Networking.java
@@ -10,6 +10,11 @@ public class Networking {
     public static void init() {
     }
 
+    /**
+     * Send a packet to clients updating the drawn pattern on a scroll.
+     * @param entity the hanging scroll to update
+     * @param level the world the scroll is in
+     */
     public static void sendScrollSync(EntityWallScroll entity, ServerLevel level) {
         ScrollSyncPacket packet = new ScrollSyncPacket(entity.getId(), entity.scroll);
 
@@ -18,8 +23,15 @@ public class Networking {
         }
     }
 
-    public static void sendMindImpression(ServerLevel level, Entity source, BlockPos inducerPos) {
-        MindImpressionPacket packet = new MindImpressionPacket(level.dimension().location(), source.getId(), inducerPos);
+    /**
+     * Sends a packet to clients telling them to draw particles for a Cognitive Inducer's mind impression.
+     * @param level the world the impression took place in
+     * @param source the entity that was impressed
+     * @param inducerPos the position of the inducer
+     * @param successful if the impression was successful or not (for particle color selection)
+     */
+    public static void sendMindImpression(ServerLevel level, Entity source, BlockPos inducerPos, boolean successful) {
+        MindImpressionPacket packet = new MindImpressionPacket(level.dimension().location(), source.getId(), inducerPos, successful);
 
         for (ServerPlayer player : level.players()) {
             packet.sendToPlayer(player);