+++ fabric/src/main/java/name/dashkal/minecraft/hexresearch/fabric/cc/DeprecatedComponent.java
@@ -1,0 +1,5 @@
+package name.dashkal.minecraft.hexresearch.fabric.cc;
+
+import dev.onyxstudios.cca.api.v3.component.ComponentV3;
+
+public interface DeprecatedComponent extends ComponentV3 {}